Migrate helpers for server_install
